Fixed Gesture Bugs, Performance Improvement, Fixed Some Crashes, and probably some other stuff?

+++ MaizeUI/MZECAPackageView.m
@@ -41,17 +41,17 @@
 }
 - (void)setStateName:(NSString *)name {
 	if (_stateController && _packageLayer) {
-		CGFloat transitionSpeed = 1.0;
+		float transitionSpeed = 1.0f;
 		if ([name containsString:@"|"]) {
 			NSArray* split = [name componentsSeparatedByString:@"|"];
 			if ([split count] == 2) {
-				transitionSpeed = (CGFloat)[(NSString *)split[1] floatValue];
+				transitionSpeed = (float)[(NSString *)split[1] floatValue];
 				name = (NSString *)split[0];
 			}
 		}
 
 		CAState *state = [_packageLayer stateWithName:name];
-		[_stateController setState:state ofLayer:_packageLayer transitionSpeed:transitionSpeed];
+		[_stateController setState:state ofLayer:_packageLayer transitionSpeed:(float)transitionSpeed];
 	}
 
 }
